feat: add healthcheck (#9)
adds an http server to the backup process that provides a healthcheck
status. We're seeing the backup process just lock up in prod with no
logging but also no process failure or other error to help debug. The
plan is to wire this in to ECS so that it restarts the container if it
fails healthcheck.

if it's been more than 2 mins since last log, assume we are stalled and
return 500.

update dockerfile to use curl to test the healthcheck url

License: MIT

---------

Signed-off-by: Oli Evans <oli@protocol.ai>